expanding coverage tests
