+++ Sources/Cachew/Common/Storable.swift
@@ -1,0 +1,12 @@
+//
+//  Storable.swift
+//  Cachew
+//
+//  Created by Lucas Migge on 13/07/25.
+//
+
+
+public protocol Storable: Sendable & Codable {
+    var cacheKey: any CachewKey { get }
+}
+
