+++ Sources/Cachew/Silo/FileManager/FileManagerProtocol.swift
@@ -19,12 +19,3 @@ public protocol FileManagerProtocol {
     func readData(from url: URL) throws -> Data
 }
 
-extension FileManager: FileManagerProtocol {
-    public func write(data: Data, to url: URL) throws {
-        try data.write(to: url, options: .atomic)
-    }
-    
-    public func readData(from url: URL) throws -> Data {
-        try Data(contentsOf: url)
-    }
-}
