--- Sources/Cachew/Silo/Storable.swift
@@ -1,9 +1,0 @@
-//
-//  Storable.swift
-//  Cachew
-//
-//  Created by Lucas Migge on 13/07/25.
-//
-
-
-public typealias Storable = Sendable & Codable
